Camera Tuning
Instead of the display being simply shut off after the round end condition is reached, the active camera is set to a wide-shot of the course. A blue/black skybox was added for aesthetic touch, and a script for the camera swapping was put in along with many comments.

Current Bugs:
-Enemy will not move at game start if it checks for the game state to be Playing. (haven't investigated this since the last version)

Features in Progress:
-Updating the camera swapping script to check for tags and add cameras to an array of cameras rather than hard coded game objects. This should make it easier to adapt for other levels and can provide cinematic shots or allow the players to change their PoV on demand.